requirements.txt
gql[all]>=0.0.dev0
pygit2
pygithub
pandas
psycopg2-binary
pytest-timeout
gitpython
oyaml